MasS-python/pyproject.toml
[project]
name = "mass-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.128.7",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "loguru>=0.7.3",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.39",
    "uvicorn[standard]>=0.40.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
